[move/#229] ContentUriRequestBody 파일 이동
