+++ data/presigned-url/src/main/java/com/napzak/market/presigned_url/datasource/PresignedUrlDataSource.kt
@@ -3,7 +3,7 @@ package com.napzak.market.presigned_url.datasource
 import android.content.ContentResolver
 import androidx.core.net.toUri
 import com.napzak.market.presigned_url.service.PresignedUrlService
-import com.napzak.market.util.android.ContentUriRequestBody
+import com.napzak.market.remote.ContentUriRequestBody
 import javax.inject.Inject
 
 class PresignedUrlDataSource @Inject constructor(
