+++ data/remote/src/main/java/com/napzak/market/remote/ContentUriRequestBody.kt
@@ -1,0 +1,50 @@
+package com.napzak.market.remote
+
+import android.content.ContentResolver
+import android.net.Uri
+import android.provider.MediaStore
+import okhttp3.MediaType
+import okhttp3.MediaType.Companion.toMediaTypeOrNull
+import okhttp3.MultipartBody
+import okhttp3.RequestBody
+import okio.BufferedSink
+import okio.source
+
+private const val IMAGE = "image"
+
+class ContentUriRequestBody(
+    private val contentResolver: ContentResolver,
+    private val uri: Uri,
+) : RequestBody() {
+    private var fileName = ""
+    private var size = -1L
+
+    init {
+        contentResolver.query(
+            uri,
+            arrayOf(MediaStore.Images.Media.SIZE, MediaStore.Images.Media.DISPLAY_NAME),
+            null,
+            null,
+            null,
+        )?.use { cursor ->
+            if (cursor.moveToFirst()) {
+                size = cursor.getLong(cursor.getColumnIndexOrThrow(MediaStore.Images.Media.SIZE))
+                fileName =
+                    cursor.getString(cursor.getColumnIndexOrThrow(MediaStore.Images.Media.DISPLAY_NAME))
+            }
+        }
+    }
+
+    private fun getFileName() = fileName
+
+    override fun contentLength(): Long = size
+
+    override fun contentType(): MediaType? = contentResolver.getType(uri)?.toMediaTypeOrNull()
+
+    override fun writeTo(sink: BufferedSink) {
+        contentResolver.openInputStream(uri)?.source()?.use { source -> sink.writeAll(source) }
+    }
+
+    fun toFormData(name: String = IMAGE) =
+        MultipartBody.Part.createFormData(name, getFileName(), this)
+}
